+++ src/mxcp/dbt/__init__.py
@@ -1,0 +1,5 @@
+"""MXCP dbt integration module.
+
+This module provides functionality for dbt integration with MXCP,
+including configuration management and wrapper commands.
+"""
